refactor(consumption): split receipt_parser into extractor modules (Refs #563)
Break the 607-LOC `receipt_parser.dart` into four cohesive files under
`receipt_parser/` so each sits under the 300-LOC guideline without
forcing any caller to edit its imports:

- `receipt_parse_result.dart` — the result data class only (45 LOC).
- `receipt_field_extractors.dart` — pure-function extractors for
  fuel type, liters, total cost, price-per-litre, date, and the
  shared `matchFirst` / `parseDecimal` / `decimalDigitCount` helpers
  (262 LOC).
- `brand_detection.dart` — `detectBrand` + `extractStationName`
  (51 LOC).
- `brand_layouts.dart` — the per-brand dispatch (`parseSuperU`,
  `parseCarrefour`, `parseGeneric`) plus `reconcile` (134 LOC).

The top-level `receipt_parser.dart` shrinks to 124 LOC — it now holds
only the `ReceiptParser` class, its override-dispatch logic, and a
re-export of `ReceiptParseResult` so existing consumers (receipt_scan
_service, bad_scan_report_sheet, the two parser test files) keep
compiling without any import changes.

Why: unblock epic #563's next phase. Pure extraction, zero behaviour
change — the 58 existing receipt parser tests pass unchanged, and the
full 5744-test suite stays green.
